+++ robot-system-lib-parent/robot-system-lib-core/src/main/java/org/arig/robot/model/lidar/ConfigurationInfos.java
@@ -12,7 +12,7 @@ import java.io.Serializable;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class ConfigurationInfos implements Serializable {
   private Boolean reversed = false;
-  private Integer angleOffset = 0;
+  private Double angleOffset = 0.0;
   private Integer excludeLowerThanMm = 150;
   private Integer excludeGreaterThanMm = 3600;
 }
