+++ robot-system-lib-parent/robot-system-lib-core/src/main/java/org/arig/robot/system/capteurs/i2c/GP2D12Telemeter.java
@@ -126,7 +126,7 @@ public class GP2D12Telemeter implements ILidarTelemeter {
   }
 
   @Override
-  public void setConfiguration(boolean reverse, int offsetAngle, int excludeLowerThan, int excludeGreaterThan) {
+  public void setConfiguration(boolean reverse, double offsetAngle, int excludeLowerThan, int excludeGreaterThan) {
     // NOOP
   }
