+++ robot-system-lib-parent/robot-system-lib-core/src/test/java/org/arig/robot/system/capteurs/LidarTelemeterMock.java
@@ -74,7 +74,7 @@ public class LidarTelemeterMock implements ILidarTelemeter {
   }
 
   @Override
-  public void setConfiguration(boolean reverse, int offsetAngle, int excludeLowerThan, int excludeGreaterThan) {
+  public void setConfiguration(boolean reverse, double offsetAngle, int excludeLowerThan, int excludeGreaterThan) {
   }
 
   @Override
